multi_agent_system/requirements.txt
# Core dependencies
rich>=13.0.0
aiohttp>=3.9.0
asyncio>=3.4.3

# Docker SDK
docker>=7.0.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Data validation
pydantic>=2.0.0

# Web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
sse-starlette>=2.0.0

# Database
aiosqlite>=0.19.0
sqlalchemy[asyncio]>=2.0.0

# Testing & E2E
sseclient>=3.1.0
requests>=2.31.0
